move intent driven API for location lookups into base and app packages
