+++ unifiednlp-base/src/main/java/org/microg/tools/selfcheck/NlpOsCompatChecks.java
@@ -17,6 +17,7 @@
 package org.microg.tools.selfcheck;
 
 import android.content.Context;
+import android.content.SharedPreferences;
 
 import org.microg.nlp.R;
 
@@ -26,7 +27,8 @@ import static android.os.Build.VERSION.SDK_INT;
 import static android.os.Build.VERSION_CODES.JELLY_BEAN;
 import static android.os.Build.VERSION_CODES.JELLY_BEAN_MR1;
 import static android.os.Build.VERSION_CODES.KITKAT;
-import static android.os.Build.VERSION_CODES.M;
+import static android.os.Build.VERSION_CODES.N_MR1;
+import org.microg.nlp.api.Constants;
 
 public class NlpOsCompatChecks implements SelfCheckGroup {
 
@@ -47,7 +49,7 @@ public class NlpOsCompatChecks implements SelfCheckGroup {
     }
 
     private boolean checkSystemIsSupported(Context context, ResultCollector collector) {
-        boolean isSupported = (SDK_INT >= KITKAT && SDK_INT <= M);
+        boolean isSupported = (SDK_INT >= KITKAT && SDK_INT <= N_MR1);
         collector.addResult(context.getString(R.string.self_check_name_system_supported),
                 isSupported ? Result.Positive : Result.Unknown, context.getString(R.string.self_check_resolution_system_supported));
         return isSupported;
@@ -74,6 +76,13 @@ public class NlpOsCompatChecks implements SelfCheckGroup {
         }
         collector.addResult(context.getString(R.string.self_check_name_nlp_package_name),
                 systemMatchesPackage ? Result.Positive : Result.Negative, context.getString(R.string.self_check_resolution_nlp_package_name));
+        
+        SharedPreferences mSharedPreferences = context.getSharedPreferences(Constants.APP_SETTINGS_NAME,
+                Context.MODE_PRIVATE);
+        SharedPreferences.Editor editor = mSharedPreferences.edit();
+        editor.putBoolean(Constants.OS_SUPPORTS_UNLP, systemMatchesPackage);
+        editor.apply();
+        
         return systemMatchesPackage;
     }
