--- unifiednlp-standalone/src/main/res/xml/nlp_injected_setting.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright 2013-2016 microG Project Team
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<injected-location-setting xmlns:android="http://schemas.android.com/apk/res/android"
-    android:title="@string/nlp_settings_label"
-    android:icon="@mipmap/ic_nlp_settings"
-    android:settingsActivity="org.microg.nlp.ui.StandaloneSettingsActivity" />
