--- unifiednlp-standalone/src/main/res/xml/nlp_preferences.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright 2013-2016 microG Project Team
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    android:key="nlp_preferences">
-    <PreferenceCategory android:title="@string/network_location">
-        <org.microg.nlp.ui.StandaloneLocationBackendPreference
-            android:key="@string/pref_location_backends"
-            android:title="@string/configure_backends"
-            android:summary="@string/configure_location_backends_summary" />
-    </PreferenceCategory>
-    <PreferenceCategory android:title="@string/geocoding">
-        <org.microg.nlp.ui.StandaloneGeocoderBackendPreference
-            android:key="@string/pref_geocoder_backends"
-            android:title="@string/configure_backends"
-            android:summary="@string/configure_geocoder_backends_summary" />
-    </PreferenceCategory>
-</PreferenceScreen>
